rename and replace of files ssooiigle (main) and threadData
